Extract new dynamic table story to a new component
