+++ packages/ui/react-ui-table/src/components/Table/index.ts
@@ -3,3 +3,4 @@
 //
 
 export * from './Table';
+export * from './DynamicTable';
